refactor: move unlock logic to a component
